+++ packages/next/client/page-loader.ts
@@ -292,7 +292,9 @@ export default class PageLoader {
                 d.endsWith('.js') &&
                 !document.querySelector(`script[src^="${d}"]`)
               ) {
-                this.loadScript(d, route)
+                import('react').then(()=>{
+                  this.loadScript(d, route)
+                })
               }
               if (
                 d.endsWith('.css') &&
